@@ -2,6 +2,7 @@
   "name": "prettier-standard",
   "version": "16.4.1",
   "description": "Prettier and standard brought together!",
+  "type": "module",
   "main": "src/index.js",
   "bin": {
     "prettier-standard": "src/cli.js"
@@ -14,34 +15,33 @@
     "formatting"
   ],
   "engines": {
-    "node": ">=8"
+    "node": ">=14"
   },
   "author": "Adam Stankiewicz <sheerun@sher.pl>",
   "license": "MIT",
   "repository": "sheerun/prettier-standard",
   "dependencies": {
-    "chalk": "^2.4.2",
-    "diff": "^4.0.2",
-    "eslint": "^6.8.0",
-    "execa": "^2.0.4",
-    "find-up": "^4.1.0",
-    "get-stdin": "^7.0.0",
-    "globby": "^6.1.0",
-    "ignore": "^3.3.7",
-    "lint-staged": "9.4.3",
-    "mri": "^1.1.5",
-    "multimatch": "^3.0.0",
-    "prettierx": "0.11.3"
+    "chalk": "^4.1.2",
+    "eslint": "^8.2.0",
+    "execa": "^6.0.0",
+    "find-up": "^6.2.0",
+    "get-stdin": "^9.0.0",
+    "globby": "^12.0.2",
+    "ignore": "^5.1.9",
+    "lint-staged": "12.0.2",
+    "mri": "^1.2.0",
+    "multimatch": "^6.0.0",
+    "prettierx": "0.19.0"
   },
   "devDependencies": {
-    "isomorphic-git": "1.3.2",
-    "jest": "25.4.0",
-    "tmp": "0.1.0"
+    "isomorphic-git": "1.10.1",
+    "jest": "27.3.1",
+    "tmp": "0.2.1"
   },
   "scripts": {
     "format": "node src/cli.js --format",
     "lint": "node src/cli.js --lint",
-    "test": "jest"
+    "test": "NODE_OPTIONS=--experimental-vm-modules jest"
   },
   "files": [
     "src"
