@@ -1,7 +1,7 @@
-const findUp = require('find-up')
-const execa = require('execa')
-const path = require('path')
-const parseDiff = require('../parse-diff')
+import { findUpSync } from 'find-up'
+import { execaSync } from 'execa'
+import path from 'path'
+import parseDiff from '../parse-diff.js'
 
 const SPECIAL_EMPTY_TREE_COMMIT_HASH =
   '4b825dc642cb6eb9a060e54bf8d69288fbee4904'
@@ -41,7 +41,7 @@ class Git {
   }
 
   runGit (args) {
-    return execa.sync('git', args, { cwd: this.cwd }).stdout.trim()
+    return execaSync('git', args, { cwd: this.cwd }).stdout.trim()
   }
 
   getRevision (branch) {
@@ -71,8 +71,8 @@ class Git {
   }
 }
 
-module.exports = cwd => {
-  const gitDirectory = findUp.sync('.git', { cwd, type: 'directory' })
+export default cwd => {
+  const gitDirectory = findUpSync('.git', { cwd, type: 'directory' })
 
   if (gitDirectory) {
     return new Git(path.dirname(gitDirectory), cwd)
